# services/src/services/shared/old_openapi.json.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API V3 - Data Etat - Budget",
    "description": "\nApi de d'accès aux données financières de l'état\n\u003Cb\u003EC'est une API dediée à l'outil interne de consultation budget. N'utilisez pas cette API pour intégrer nos données à votre système.\u003C/b\u003E\n    ",
    "version": "3.0"
  },
  "servers": [
    {
      "url": "/financial-data/api/v3"
    }
  ],
  "paths": {
...
    "/lignes": {
      "get": {
        "tags": [
          "Lignes financieres"
        ],
        "summary": "Récupére les lignes financières, mécanisme de grouping pour récupérer les montants agrégés",
        "operationId": "get_lignes_financieres_lignes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "force_no_cache",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force No Cache"
            }
          },
          {
            "name": "source_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Region"
            }
          },
          {
            "name": "data_source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Source"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "n_ej",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "N Ej"
            }
          },
          {
            "name": "code_programme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code Programme"
            }
          },
          {
            "name": "niveau_geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Niveau Geo"
            }
          },
          {
            "name": "code_geo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code Geo"
            }
          },
          {
            "name": "ref_qpv",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [2015, 2024],
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ref Qpv"
            }
          },
          {
            "name": "code_qpv",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code Qpv"
            }
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          },
          {
            "name": "beneficiaire_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Siret du bénéficiaire",
              "title": "Beneficiaire Code"
            },
            "description": "Siret du bénéficiaire"
          },
          {
            "name": "beneficiaire_categorieJuridique_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Type de la catégorie juridique du bénéficiaire",
              "title": "Beneficiaire Categoriejuridique Type"
            },
            "description": "Type de la catégorie juridique du bénéficiaire"
          },
          {
            "name": "annee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annee"
            }
          },
          {
            "name": "centres_couts",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Centres Couts"
            }
          },
          {
            "name": "domaine_fonctionnel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domaine Fonctionnel"
            }
          },
          {
            "name": "referentiel_programmation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Referentiel Programmation"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "grouping",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grouping"
            }
          },
          {
            "name": "grouped",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grouped"
            }
          },
          {
            "name": "colonnes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Liste des codes des colonnes à récupérer, séparés par des virgules",
              "title": "Colonnes"
            },
            "description": "Liste des codes des colonnes à récupérer, séparés par des virgules"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "asc",
                    "desc"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "fields_search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fields Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LignesResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": 422,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": 500,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lignes/{id}": {
      "get": {
        "tags": [
          "Lignes financieres"
        ],
        "summary": "Récupére les infos budgetaires en fonction de son identifiant technique",
        "operationId": "get_lignes_financieres_by_source_lignes__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "source_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Region"
            }
          },
          {
            "name": "data_source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Source"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "colonnes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colonnes"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "asc",
                    "desc"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort Order"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "fields_search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fields Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LigneResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": 422,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": 500,
                  "success": false,
                  "timestamp": "1764235966"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "success": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "APIError",
        "example": {
          "code": 422,
          "detail": "",
          "message": "Erreur de validation",
          "success": false,
          "timestamp": "2025-11-27T10:32:45.854166+01:00"
        }
      },
...
      "GroupedData": {
        "properties": {
          "colonne": {
            "type": "string",
            "title": "Colonne",
            "default": ""
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_montant_engage": {
            "type": "number",
            "title": "Total Montant Engage",
            "default": 0
          },
          "total_montant_paye": {
            "type": "number",
            "title": "Total Montant Paye",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "label",
          "value",
          "total"
        ],
        "title": "GroupedData"
      },
      "Groupings": {
        "properties": {
          "type": {
            "type": "string",
            "const": "groupings",
            "title": "Type",
            "default": "groupings"
          },
          "total": {
            "$ref": "#/components/schemas/Total"
          },
          "groupings": {
            "items": {
              "$ref": "#/components/schemas/GroupedData"
            },
            "type": "array",
            "title": "Groupings"
          }
        },
        "type": "object",
        "required": [
          "total",
          "groupings"
        ],
        "title": "Groupings"
      },
...
      "LignesResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "success": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LignesFinancieres"
              },
              {
                "$ref": "#/components/schemas/Groupings"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "pagination"
        ],
        "title": "LignesResponse"
      },
      "PaginationMeta": {
        "properties": {
          "current_page": {
            "type": "integer",
            "title": "Current Page"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "current_page",
          "has_next"
        ],
        "title": "PaginationMeta"
      },
...
      "Total": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "total_montant_engage": {
            "type": "number",
            "title": "Total Montant Engage",
            "default": 0
          },
          "total_montant_paye": {
            "type": "number",
            "title": "Total Montant Paye",
            "default": 0
          }
        },
        "type": "object",
        "title": "Total"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "http://localhost:8080/realms/Regions/protocol/openid-connect/token"
          }
        }
      }
    }
  }
}